remove print statement, clean up comments, and create test_config.py
