+++ tests/test_freezable.py
@@ -4,7 +4,7 @@ from __future__ import annotations  # types are strings by default in 3.11
 import unittest
 from dataclasses import dataclass
 
-from elfpy.types import freezable, Config
+from elfpy.types import freezable
 
 
 @dataclass
@@ -52,7 +52,7 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_not_set(self):
         """
-        freezable object when not frozen is NOT set, and no_new_attribs is NOT set
+        freezable object when frozen is NOT set, and no_new_attribs is NOT set
         desired behavior is: CAN change attributes, and CAN add attributes
         """
         freezable_object = FreezableClass()
@@ -61,8 +61,8 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_frozen_is_set(self):
         """
-        freezable object when not frozen IS set, but no_new_attribs is NOT set
-        desired behavior is: can NOT change attributes, but CAN add attributes
+        freezable object when frozen IS set, but no_new_attribs is NOT set, manually
+        desired behavior is: can NOT change attributes, and CAN add attributes
         """
         freezable_object = FreezableClass()
         freezable_object.freeze()  # pylint: disable=no-member # type: ignore
@@ -72,8 +72,8 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_frozen_is_set_by_default(self):
         """
-        freezable object when not frozen IS set, but no_new_attribs is NOT set
-        desired behavior is: can NOT change attributes, but CAN add attributes
+        freezable object when frozen IS set, but no_new_attribs is NOT set, by default
+        desired behavior is: can NOT change attributes, and CAN add attributes
         """
         freezable_object = FreezableClassFrozenByDefault()
         with self.assertRaises(AttributeError):
@@ -82,8 +82,8 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_no_new_attribs_is_set(self):
         """
-        freezable object when not frozen is NOT set, but no_new_attribs IS set
-        desired behavior is: CAN change attributes, but can NOT add attributes
+        freezable object when frozen is NOT set, but no_new_attribs IS set, manually
+        desired behavior is: CAN change attributes, and can NOT add attributes
         """
         freezable_object = FreezableClass()
         freezable_object.disable_new_attribs()  # pylint: disable=no-member # type: ignore
@@ -93,8 +93,8 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_no_new_attribs_is_set_by_default(self):
         """
-        freezable object when not frozen is NOT set, but no_new_attribs IS set
-        desired behavior is: CAN change attributes, but can NOT add attributes
+        freezable object when frozen is NOT set, but no_new_attribs IS set, by default
+        desired behavior is: CAN change attributes, and can NOT add attributes
         """
         freezable_object = FreezableClassNoNewAttribsByDefault()
         freezable_object.existing_attrib = 2
@@ -103,8 +103,8 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_frozen_is_set_and_no_new_attribs_is_set(self):
         """
-        freezable object when not frozen is NOT set, but no_new_attribs IS set
-        desired behavior is: CAN change attributes, but can NOT add attributes
+        freezable object when frozen IS set, and no_new_attribs IS set, manually
+        desired behavior is: can NOT change attributes, and can NOT add attributes
         """
         freezable_object = FreezableClass()
         freezable_object.freeze()  # pylint: disable=no-member # type: ignore
@@ -116,19 +116,11 @@ class TestFreezability(unittest.TestCase):
 
     def test_freezable_when_frozen_is_set_and_no_new_attribs_is_set_by_default(self):
         """
-        freezable object when not frozen is NOT set, but no_new_attribs IS set
-        desired behavior is: CAN change attributes, but can NOT add attributes
+        freezable object when frozen IS set, and no_new_attribs IS set, by default
+        desired behavior is: can NOT change attributes, and can NOT add attributes
         """
         freezable_object = FreezableClassFrozenNoNewAttribsByDefault()
         with self.assertRaises(AttributeError):
             freezable_object.existing_attrib = 2
         with self.assertRaises(AttributeError):
             freezable_object.new_attrib = 1  # pylint: disable=attribute-defined-outside-init # type: ignore
-
-    def test_config_cant_add_new_attribs(self):
-        """
-        config object can't add new attributes after it's initialized
-        """
-        config = Config()
-        with self.assertRaises(AttributeError):
-            config.new_attrib = 1
